+++ ios/Video/RCTVideoManager.h
@@ -1,5 +1,6 @@
 #import <React/RCTViewManager.h>
+#import <React/RCTBridgeModule.h>
 
-@interface RCTVideoManager : RCTViewManager
+@interface RCTVideoManager : RCTViewManager <RCTBridgeModule>
 
 @end